# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SBG Internal Portal API",
    "description": "REST API for the AWS Student Builder Group internal operations portal — members, tasks, submissions, link shortener, audit logs, and analytics. All endpoints (except auth) require a valid `sbg_session` HttpOnly cookie obtained via the OTP login flow.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Relative to deployment origin" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sbg_session"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-secret",
        "description": "Shared secret configured via the CRON_SECRET env var. An `Authorization: Bearer <secret>` header is also accepted."
      }
    },
    "schemas": {
...
      "Domain": {
        "type": "string",
        "enum": ["Technical", "Corporate", "Creatives", "General"]
      },
      "Subdomain": {
        "type": "string",
        "enum": ["Software Development", "AI & Machine Learning", "Cloud & DevOps", "Events & Operations", "Sponsorship & Finance", "HR & Admin", "PR & Marketing", "Digital Design", "Media Production"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "taskId": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "assignmentType": { "type": "string", "enum": ["INDIVIDUAL", "DOMAIN", "SUBDOMAIN", "GENERAL"] },
          "assignedToId": { "type": "string", "nullable": true },
          "assignedToName": { "type": "string" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "createdBy": { "type": "string" },
          "createdByName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"], "description": "Lazily flipped OPEN->CLOSED on read once the deadline has passed — no scheduled job required." },
          "totalSubmissions": { "type": "integer" },
          "reminderSentAt": { "type": "string", "format": "date-time", "nullable": true, "description": "Set once a deadline-reminder email batch has been sent for this task; cleared if the deadline is changed." }
        }
      },
      "TaskPriority": {
        "type": "string",
        "enum": ["LOW", "MEDIUM", "HIGH"]
      },
      "TaskCreateRequest": {
        "type": "object",
        "required": ["title", "description", "deadline", "assignmentType"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time" },
          "priority": { "$ref": "#/components/schemas/TaskPriority", "description": "Defaults to MEDIUM if omitted" },
          "assignmentType": { "type": "string", "enum": ["INDIVIDUAL", "DOMAIN", "SUBDOMAIN", "GENERAL"] },
          "assignedToId": { "type": "string", "description": "memberId, required when assignmentType=INDIVIDUAL" },
          "domain": { "$ref": "#/components/schemas/Domain" },
          "subdomain": { "$ref": "#/components/schemas/Subdomain" }
        }
      },
      "TaskUpdateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date-time", "description": "Changing this clears reminderSentAt so a new reminder cycle can fire" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"] }
        }
      },
      "TaskDetail": {
        "type": "object",
        "properties": {
          "task": { "$ref": "#/components/schemas/Task" },
          "submissions": { "type": "array", "items": { "$ref": "#/components/schemas/Submission" } },
          "mySubmission": { "allOf": [{ "$ref": "#/components/schemas/Submission" }], "nullable": true },
          "canReview": { "type": "boolean" },
          "canSubmit": { "type": "boolean" },
          "canDelete": { "type": "boolean" },
          "canClose": { "type": "boolean" }
        }
      },
      "Submission": {
        "type": "object",
        "properties": {
          "submissionId": { "type": "string" },
          "taskId": { "type": "string" },
          "taskTitle": { "type": "string" },
          "memberId": { "type": "string" },
          "memberName": { "type": "string" },
          "content": { "type": "string" },
          "links": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "submittedAt": { "type": "string", "format": "date-time" },
          "reviewStatus": { "type": "string", "enum": ["PENDING", "APPROVED", "REJECTED"] },
          "reviewedBy": { "type": "string", "nullable": true },
          "reviewedByName": { "type": "string", "nullable": true },
          "reviewedAt": { "type": "string", "format": "date-time", "nullable": true },
          "reviewFeedback": { "type": "string", "nullable": true, "description": "Optional comment left by the reviewer on APPROVE/REJECT" },
          "ratingAwarded": { "type": "number", "nullable": true },
          "deadline": { "type": "string", "format": "date-time" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true }
        }
      },
      "SubmissionCreateRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": { "type": "string" },
          "links": { "type": "array", "items": { "type": "string", "format": "uri" } }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid session cookie present",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Unauthorized" } } }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted to perform this action",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Forbidden" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List tasks visible to the caller",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["OPEN", "CLOSED"] } },
          { "name": "my", "in": "query", "schema": { "type": "boolean" }, "description": "Only tasks assigned to the caller or assignmentType=GENERAL" }
        ],
        "responses": {
          "200": { "description": "Task list", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a task",
        "description": "Requires task-creation privileges (presidium, director, manager, or associate, scope-dependent). The assignment target (member/domain/subdomain) is permission-checked server-side; assignedToName is always derived server-side.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskCreateRequest" } } } },
        "responses": {
          "200": { "description": "Task created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Task" } } } } } },
          "400": { "description": "Missing required fields" },
          "403": { "description": "Unauthorized to create tasks, or not allowed to assign to this target" },
          "404": { "description": "Target member not found" }
        }
      }
    },
    "/tasks/{taskId}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task detail with submissions and caller permissions",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task detail", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/TaskDetail" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Task not found or not visible to caller" }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task (title, description, deadline, status)",
        "description": "Only the task creator or presidium may edit. Reopening a CLOSED task requires reviewer privileges (presidium/director/manager).",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskUpdateRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid status value, or no valid fields to update" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task and cascade-delete its submissions",
        "description": "Only the task creator or presidium may delete.",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/tasks/{taskId}/submit": {
      "post": {
        "tags": ["Submissions"],
        "summary": "Submit work for a task",
        "description": "Caller must be eligible for the task's assignment scope, and the task must be OPEN (a task past its deadline is auto-closed on this read). A prior REJECTED submission does not block a new attempt — only an existing PENDING or APPROVED submission does.",
        "parameters": [{ "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubmissionCreateRequest" } } } },
        "responses": {
          "200": { "description": "Submission created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Submission" } } } } } },
          "400": { "description": "Task is closed, or content is required" },
          "403": { "description": "Not eligible to submit to this task" },
          "404": { "description": "Task not found" },
          "409": { "description": "Already submitted for this task" }
        }
      }
    },
...
  }
}